fix: v2 audit: https://github.com/yAudit/temple-lending-report/issues/35

+++ protocol/contracts/amo/Ramos.sol
@@ -74,7 +74,7 @@ contract Ramos is IRamos, TempleElevatedAccess, Pausable {
     /// @notice The Treasury Price Index (TPI) Oracle
     ITreasuryPriceIndexOracle public override tpiOracle;
 
-    /// @notice The vault from where to borrow and repay the Protocol Token
+    /// @notice The vault from where to borrow and repay the Protocol & Quote Tokens
     IRamosTokenVault public override tokenVault;
 
     /// @notice The percentage bounds (in bps) beyond which to rebalance up or down
@@ -184,7 +184,7 @@ contract Ramos is IRamos, TempleElevatedAccess, Pausable {
     }
 
     /**
-     * @notice Set the Treasury Price Index (TPI) Oracle
+     * @notice Set the token vault - where to borrow and repay the Protocol & Quote Tokens
      */
     function setTokenVault(address vault) external override onlyElevatedAccess {
         emit TokenVaultSet(vault);
@@ -237,7 +237,7 @@ contract Ramos is IRamos, TempleElevatedAccess, Pausable {
     /**
      * @notice The Treasury Price Index - the target price of the Treasury, in `quoteToken` terms.
      */
-    function treasuryPriceIndex() public view override returns (uint256) {
+    function treasuryPriceIndex() public view override returns (uint96) {
         return tpiOracle.treasuryPriceIndex();
     }
